+++ borax/choices.py
@@ -1,7 +1,7 @@
 # coding=utf8
-
-
 from collections import OrderedDict
+
+from typing import Dict
 
 __all__ = ['Item', 'ConstChoices']
 
@@ -10,30 +10,44 @@ class Item:
     _order = 0
 
     def __init__(self, value, display=None, *, order=None):
-        self.value = value
+        self._value = value
         if display is None:
-            self.display = str(value)
+            self._display = str(value)
         else:
-            self.display = str(display)
+            self._display = str(display)
         if order is None:
             Item._order += 1
             self.order = Item._order
         else:
             self.order = order
 
+    @property
+    def value(self):
+        return self._value
+
+    @property
+    def display(self):
+        return self._display
+
+    @property
+    def label(self):
+        return self._display
+
+    def __str__(self):
+        return '<{0} value={1!r} label={2!r}>'.format(self.__class__.__name__, self.value, self.label)
+
+    __repr__ = __str__
+
 
 class ChoicesMetaclass(type):
     def __new__(cls, name, bases, attrs):
 
-        choices = []
-        display_lookup = {}
-
-        fields = {}
+        fields = {}  # {<name>:<BItem>}
 
         parents = [b for b in bases if isinstance(b, ChoicesMetaclass)]
         for kls in parents:
-            for field_name in kls._fields:
-                fields[field_name] = kls._fields[field_name]
+            for field_name in kls.fields:
+                fields[field_name] = kls.fields[field_name]
 
         for k, v in attrs.items():
             if k.startswith('_'):
@@ -41,34 +55,48 @@ class ChoicesMetaclass(type):
             if isinstance(v, Item):
                 fields[k] = v
             elif isinstance(v, (tuple, list)) and len(v) == 2:
-                fields[k] = Item(value=v[0], display=v[1])
+                fields[k] = Item(v[0], v[1])
             elif isinstance(v, (int, float, str, bytes)):
-                fields[k] = Item(value=v, display=v)
+                fields[k] = Item(v)
 
         fields = OrderedDict(sorted(fields.items(), key=lambda x: x[1].order))
+        for field_name, item in fields.items():
+            attrs[field_name] = item.value  # override the exists attrs __dict__
 
-        for field_name in fields:
-            val_item = fields[field_name]
-
-            if isinstance(val_item, Item):
-                choices.append((val_item.value, val_item.display))
-                display_lookup[val_item.value] = val_item.display
-                attrs[field_name] = val_item.value
-            else:
-                choices.append((field_name, val_item.choices))
-
-        attrs['_fields'] = fields
-        attrs['_choices'] = choices
-        attrs['_display_lookup'] = display_lookup
-        return type.__new__(cls, name, bases, attrs)
+        new_cls = super().__new__(cls, name, bases, attrs)
+        new_cls._fields = fields
+        return new_cls
 
     @property
-    def choices(self):
-        return self._choices
+    def fields(cls) -> Dict[str, Item]:
+        return dict(cls._fields)
 
     @property
-    def display_lookup(self):
-        return self._display_lookup
+    def choices(cls) -> list:
+        return [(item.value, item.label) for _, item in cls.fields.items()]
+
+    @property
+    def names(cls) -> tuple:
+        return tuple(cls.fields.keys())
+
+    @property
+    def values(cls) -> tuple:
+        return tuple([value for value, _ in cls.choices])
+
+    @property
+    def displays(cls) -> tuple:
+        return tuple([display for _, display in cls.choices])
+
+    @property
+    def labels(cls) -> tuple:
+        return cls.displays
+
+    @property
+    def display_lookup(cls) -> dict:
+        return {value: label for value, label in cls.choices}
+
+    def __contains__(self, item):
+        return item in self.values
 
     def __iter__(self):
         for item in self.choices:
@@ -77,12 +105,13 @@ class ChoicesMetaclass(type):
     def __len__(self):
         return len(self.choices)
 
-
-class ConstChoices(metaclass=ChoicesMetaclass):
-    @classmethod
-    def is_valid(cls, value):
+    # API
+    def is_valid(cls, value) -> bool:
         return value in cls.display_lookup
 
-    @classmethod
     def get_value_display(cls, value, default=None):
         return cls.display_lookup.get(value, default)
+
+
+class ConstChoices(metaclass=ChoicesMetaclass):
+    pass