+++ borax/datasets/__init__.py
@@ -1,5 +1,1 @@
 # coding=utf8
-
-from borax.datasets.dict_datasets import DictDataset
-
-__all__ = ['DictDataset']